refactor: split explorer/tcl_cli.py into verbs/ subpackage using @verb decorator
Breaks the 2,891-line module into focused submodules:

  verbs/_registry.py   — @verb decorator + apply_verb_registrations() + get_verb_catalogue()
  verbs/_utils.py      — shared utilities: input resolution, output writing,
                         config resolution, argparse helpers, syntax highlighting
  verbs/transform.py   — opt, format, minify, unminify-error
  verbs/diag.py        — diag, lint, validate
  verbs/graphs.py      — symbols, diagram, callgraph, symbolgraph, dataflow
  verbs/lookup.py      — event-order, event-info, command-info, help
  verbs/highlight.py   — highlight
  verbs/compile.py     — dis, compwasm
  verbs/diff.py        — diff
  verbs/misc.py        — explore, convert

tcl_cli.py shrinks to ~280 lines: config loading, _BriefHelpAction/_FullHelpAction
(now driven by get_verb_catalogue() instead of a hardcoded list), parse_args, main.

_VERB_CATALOGUE is eliminated — the registry derives it from registered verbs.
pkg/venv/docker verb groups are unchanged (they use the add_*_subparser pattern
for their sub-sub-command structure).

_load_help_queries is re-exported from tcl_cli.py so existing test monkeypatches
continue to work; _run_help looks it up through the tcl_cli module reference at
call time so the patch takes effect.
